+++ src/features/map/components/MapSubPanel/MapSubPanel.module.scss
