+++ src/features/map/components/MapSubPanelItem/MapSubPanelItem.module.scss
