+++ src/features/map/components/MapSubPanelItem/MapSubPanelItem.tsx
@@ -1,4 +1,5 @@
-import { MapItemStatusFilter } from './constants/MapItemStatusFilter';
+import { MapItemStatusFilter } from '../../constants/MapItemStatusFilter';
+
 import styles from './MapSubPanelItem.module.scss';
 
 interface MapSubPanelItemProps {
